+++ src/main/kotlin/company/bigger/web/controller/UserController.kt
@@ -2,7 +2,7 @@ package company.bigger.web.controller
 
 import company.bigger.dto.UserLoginModel
 import company.bigger.dto.UserLoginModelResponse
-import company.bigger.web.jwt.ApiKeySecuredAspect
+import company.bigger.web.jwt.SecuredApi
 import org.compiere.model.I_AD_User
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.GetMapping
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import org.springframework.web.bind.annotation.RequestHeader
 import company.bigger.service.UserService
+import org.springframework.http.ResponseEntity
 
 // import org.springframework.web.bind.annotation.*
 
@@ -18,6 +19,8 @@ import company.bigger.service.UserService
 open class UserController {
     @Autowired
     private lateinit var userService: UserService
+    @Autowired
+    private lateinit var securedApi: SecuredApi
 
     @GetMapping()
     @RequestMapping(value = ["/user/{username}/login/{password}"])
@@ -27,7 +30,7 @@ open class UserController {
 
     @GetMapping()
     @RequestMapping(value = ["/user/me"])
-    fun me(@RequestHeader(value = "Authorization") authorization: String): I_AD_User? {
-        return ApiKeySecuredAspect.processAuthorization(authorization, userService, {}, { userService.currentUser() }) as I_AD_User?
+    fun me(@RequestHeader(value = "Authorization") authorization: String): ResponseEntity<I_AD_User?> {
+        return securedApi.processAuthorization(authorization) { userService.currentUser() }
     }
 }
